+++ grafana-plugin/src/utils/utils.ts
@@ -118,14 +118,3 @@ function isFieldEmpty(value: any): boolean {
 export const allFieldsEmpty = (obj: any) => every(obj, isFieldEmpty);
 
 export const isMobile = window.matchMedia('(max-width: 768px)').matches;
-
-export const getProcessEnvVarSafely = (name: string) => {
-  try {
-    return process.env[name];
-  } catch (error) {
-    console.error(error);
-    return undefined;
-  }
-};
-
-export const getIsDevelopmentEnv = () => getProcessEnvVarSafely['NODE_ENV'] === 'development';
